Rename _platform.py to platform_utils.py
