+++ source/modules/shortcut.py
@@ -5,8 +5,8 @@ import sys
 from pathlib import Path
 from shutil import copyfile
 
-from modules._platform import get_cache_path, get_cwd, get_platform, is_frozen
 from modules.icons import get_bl_file_location
+from modules.platform_utils import get_cache_path, get_cwd, get_platform, is_frozen
 from modules.settings import get_library_folder
