+++ source/threads/scraper.py
@@ -4,8 +4,8 @@ import logging
 from itertools import chain
 from typing import TYPE_CHECKING
 
-from modules._platform import get_architecture, get_platform
 from modules.build_info import BuildInfo
+from modules.platform_utils import get_architecture, get_platform
 from modules.settings import (
     get_scrape_bfa_builds,
     get_scrape_daily_builds,
