+++ source/threads/scraping/automated.py
@@ -5,8 +5,8 @@ import logging
 from datetime import UTC, datetime
 from typing import TYPE_CHECKING
 
-from modules._platform import get_architecture, get_platform
 from modules.build_info import BuildInfo, parse_blender_ver
+from modules.platform_utils import get_architecture, get_platform
 from threads.scraping.base import BuildScraper, regex_filter
 
 if TYPE_CHECKING:
