+++ source/threads/scraping/base.py
@@ -5,7 +5,7 @@ from abc import ABC, abstractmethod
 from functools import lru_cache
 from typing import TYPE_CHECKING
 
-from modules._platform import get_platform
+from modules.platform_utils import get_platform
 
 if TYPE_CHECKING:
     from collections.abc import Generator
