+++ source/threads/scraping/stable.py
@@ -11,8 +11,8 @@ from urllib.parse import urljoin
 
 import dateparser
 from bs4 import BeautifulSoup, SoupStrainer
-from modules._platform import get_architecture, get_platform, stable_cache_path
 from modules.build_info import BuildInfo, parse_blender_ver
+from modules.platform_utils import get_architecture, get_platform, stable_cache_path
 from modules.scraper_cache import ScraperCache
 from modules.settings import get_minimum_blender_stable_version
 from semver import Version
