+++ source/widgets/onboarding_setup/wizard_pages.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from modules._platform import find_app_bundle, get_platform, is_frozen
+from modules.platform_utils import find_app_bundle, get_platform, is_frozen
 from modules.settings import (
     get_actual_library_folder,
     get_actual_library_folder_no_fallback,
